farmart-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
